WIP Checkin: break out canvas classes to a scenegraph node type to remove duplication of code between BioMap and Correspondence Map drawing
