+++ src/canvas/BioMap.js
@@ -1,28 +1,20 @@
 /**
   * BioMap
-  * Mithril component representing a Biological Map with a html5 canvas element.
+  *
+  * SceneGraphNodeCanvas representing a biological map and its associated tracks
+  *
   */
-import m from 'mithril';
-import PubSub from 'pubsub-js';
-import Hammer from 'hammerjs';
-import {mix} from '../../mixwith.js/src/mixwith';
 
 import {Bounds} from '../model/Bounds';
 import {FeatureMark} from './FeatureMark';
 import {MapBackbone} from './MapBackbone';
-import {SceneGraphNodeBase} from './SceneGraphNodeBase';
+import {SceneGraphNodeCanvas} from './SceneGraphNodeCanvas';
 import { Group }  from './SceneGraphNodeGroup';
-import {DrawLazilyMixin} from './DrawLazilyMixin';
-import {RegisterComponentMixin} from '../ui/RegisterComponentMixin';
-import {selectedMap} from '../topics';
 
-export class BioMap
-       extends mix(SceneGraphNodeBase)
-       .with(DrawLazilyMixin, RegisterComponentMixin) {
+export class BioMap extends SceneGraphNodeCanvas {
 
   constructor({bioMapModel, appState, layoutBounds}) {
-    super({});
-    this.model = bioMapModel;
+    super({model:bioMapModel});
     this.model.visible = {
       start: this.model.coordinates.start,
       stop: this.model.coordinates.stop
@@ -43,144 +35,43 @@ export class BioMap
     this._layout(layoutBounds);
   }
 
- // override the children prop. getter
- // get children() {
- //   return [this.backbone].concat(
- //     this.featureMarks,
- //     this.featureLabels
- //   );
- // }
- // set children(ignore) {}
-
-  get selected() {
-    return this.appState.selection.bioMaps.indexOf(this) !== -1;
-  }
-
   /**
-   * mithril lifecycle method
+   * 
+   * Re-implement lifecycle/gestrue components as needed to appease 
+   * the items on the canvas. 
+   *
    */
-  oncreate(vnode) {
-    super.oncreate(vnode);
-    this.canvas = this.el = vnode.dom;
-    this.context2d = this.canvas.getContext('2d');
-    this.drawLazily(this.domBounds);
-  }
 
-  /**
-   * mithril lifecycle method
-   */
-  onupdate(vnode) {
-    // TODO: remove this development assistive method
-    console.assert(this.el === vnode.dom);
-    let b = new Bounds(this.el.getBoundingClientRect());
-    console.log('BioMap.onupdate', b.width, b.height, this.el);
-  }
 
-  /**
-   * mithril component render method
-   */
-  view() {
-    // store these bounds, for checking in drawLazily()
-    if(this.domBounds && ! this.domBounds.isEmptyArea) {
-      this.lastDrawnMithrilBounds = this.domBounds;
-    }
-    let b = this.domBounds || {};
-    let selectedClass = this.selected ? 'selected' : '';
-    return m('canvas', {
-      class: `cmap-canvas cmap-biomap ${selectedClass}`,
-      style: `left: ${b.left}px; top: ${b.top}px;
-              width: ${b.width}px; height: ${b.height}px;
-              transform: rotate(${this.rotation}deg);`,
-      width: b.width,
-      height: b.height
-    });
-  }
-
-  /**
-   * draw our scenegraph children our canvas element
-   */
-  draw() {
-    let ctx = this.context2d;
-    if(! ctx) return;
-    if(! this.domBounds) return;
-    console.log('BioMap canvas draw', this.domBounds.width, this.domBounds.height);
-    ctx.clearRect(0, 0, this.canvas.width, this.canvas.height);
-    ctx.save();
-    //ctx.translate(0.5, 0.5); // prevent subpixel rendering of 1px lines
-    this.visible.map(child => child && child.data.draw(ctx));
-    ctx.restore();
-    // store these bounds, for checking in drawLazily()
-    this.lastDrawnCanvasBounds = this.bounds;
-  }
-
-  /**
-   * custom gesture event dispatch listener; see LayoutContainer
-   */
-  handleGesture(evt) {
-    console.log(evt);
-    if(evt.type.match(this._gestureRegex.tap)) {
-      return this._onTap(evt);
-    }
-    else if (evt.type.match(this._gestureRegex.pinch)) {
-      return this._onZoom(evt);
-    }
-    else if(evt.type.match(this._gestureRegex.wheel)) {
-      return this._onZoom(evt);
-    }
-    else if(evt.type.match(this._gestureRegex.pan)) {
-      return this._onPan(evt);
-    }
-    return false; // dont stop evt propagation
-  }
-
-  _onTap(evt) {
-    console.log('tap');
-    console.log(evt);
-    let sel = this.appState.selection.bioMaps;
-    let i = sel.indexOf(this);
-    if(i === -1) {
-      sel.push(this);
-    }
-    else {
-      sel.splice(i, 1);
-    }
-    m.redraw();
-    console.log(this.locMap.all());
-    this.markerGroup.locMap.search(
-          {minX: evt.srcEvent.layerX,
-           maxX: evt.srcEvent.layerX,
-           minY: evt.srcEvent.layerY-5,
-           maxY: evt.srcEvent.layerY+5
-          }).forEach(hit => { console.log( hit.data)});
-    PubSub.publish(selectedMap, {
-      evt: evt,
-      data: this.appState.selection.bioMaps
-    });
-    return true;
-  }
-
-  _onZoom(evt) {
-    // TODO: send zoom event to the scenegraph elements which compose the biomap
-    // (dont scale the canvas element itself)
-    console.warn('BioMap -> onZoom -- implement me', evt);
-    this.verticalScale += evt.deltaY;
-    return true; // stop event propagation
-  }
-
-  _onPan(evt) {
-    // TODO: send pan events to the scenegraph elements which compose the biomap
-    // (dont scale the canvas element itself)
-    if(evt.direction & Hammer.DIRECTION_VERTICAL) {
-      console.warn('BioMap -> onPan -- vertically; implement me', evt);
-      return true; // stop event propagation
-    }
-    return false; // do not stop propagation
-  }
-
+//  _onTap(evt) {
+//    console.log('tap');
+//    console.log(evt);
+//    let sel = this.appState.selection.bioMaps;
+//    let i = sel.indexOf(this);
+//    if(i === -1) {
+//      sel.push(this);
+//    }
+//    else {
+//      sel.splice(i, 1);
+//    }
+//    m.redraw();
+//    console.log(this.locMap.all());
+//    this.markerGroup.locMap.search(
+//      {minX: evt.srcEvent.layerX,
+//       maxX: evt.srcEvent.layerX,
+//       minY: evt.srcEvent.layerY-5,
+//       maxY: evt.srcEvent.layerY+5
+//      }).forEach(hit => { console.log( hit.data)});
+//
+//    PubSub.publish(selectedMap, {
+//      evt: evt,
+//      data: this.appState.selection.bioMaps
+//    });
+//    return true;
+//  }
   /**
    * perform layout of backbone, feature markers, and feature labels.
    */
-
   /**
    *
    * Lifecycle of a layout change
@@ -250,6 +141,5 @@ export class BioMap
 
     markerGroup.locMap.load(fmData);
     this.locMap.load(this.visible);
-    console.log(this.locMap.all());
   }
 }
